+++ entity-view/impl/src/main/java/com/blazebit/persistence/view/impl/UpdatableExpressionVisitor.java
@@ -74,7 +74,7 @@ public class UpdatableExpressionVisitor extends PathTargetResolvingExpressionVis
     private final boolean updatable;
 
     public UpdatableExpressionVisitor(EntityMetamodel metamodel, Class<?> startClass, boolean updatable, Map<String, Type<?>> rootTypes) {
-        super(metamodel, metamodel.type(startClass), null, rootTypes);
+        super(metamodel, metamodel.type(startClass), null, null, rootTypes);
         this.updatable = updatable;
     }