+++ entity-view/testsuite/src/test/java/com/blazebit/persistence/view/testsuite/basic/EntityViewSettingTest.java
@@ -19,6 +19,14 @@ package com.blazebit.persistence.view.testsuite.basic;
 import com.blazebit.persistence.CriteriaBuilder;
 import com.blazebit.persistence.PagedList;
 import com.blazebit.persistence.PaginatedCriteriaBuilder;
+import com.blazebit.persistence.testsuite.base.jpa.category.NoDatanucleus;
+import com.blazebit.persistence.testsuite.base.jpa.category.NoEclipselink;
+import com.blazebit.persistence.testsuite.base.jpa.category.NoHibernate42;
+import com.blazebit.persistence.testsuite.base.jpa.category.NoHibernate43;
+import com.blazebit.persistence.testsuite.base.jpa.category.NoHibernate50;
+import com.blazebit.persistence.testsuite.base.jpa.category.NoHibernate51;
+import com.blazebit.persistence.testsuite.base.jpa.category.NoMySQLOld;
+import com.blazebit.persistence.testsuite.base.jpa.category.NoOpenJPA;
 import com.blazebit.persistence.testsuite.entity.Document;
 import com.blazebit.persistence.testsuite.entity.Person;
 import com.blazebit.persistence.testsuite.tx.TxVoidWork;
@@ -31,8 +39,10 @@ import com.blazebit.persistence.view.testsuite.AbstractEntityViewTest;
 import com.blazebit.persistence.view.testsuite.basic.model.CustomRootPersonView;
 import com.blazebit.persistence.view.testsuite.basic.model.DocumentWithEntityView;
 import com.blazebit.persistence.view.testsuite.basic.model.FilteredDocument;
+import com.blazebit.persistence.view.testsuite.basic.model.PersonView;
 import org.junit.Assert;
 import org.junit.Test;
+import org.junit.experimental.categories.Category;
 
 import javax.persistence.EntityManager;
 import java.util.List;
@@ -59,6 +69,9 @@ public class EntityViewSettingTest extends AbstractEntityViewTest {
                 Person o1 = new Person("pers1");
                 Person o2 = new Person("pers2");
                 Person o3 = new Person("pers3");
+                Person o4 = new Person("pers4");
+                Person o5 = new Person("pers5");
+                Person o6 = new Person("pers6");
                 o1.getLocalized().put(1, "localized1");
                 o2.getLocalized().put(1, "localized2");
                 o1.setPartnerDocument(doc1);
@@ -71,6 +84,9 @@ public class EntityViewSettingTest extends AbstractEntityViewTest {
                 doc1.getContacts().put(1, o1);
                 doc2.getContacts().put(1, o2);
                 doc3.getContacts().put(1, o3);
+                doc1.getContacts().put(2, o4);
+                doc2.getContacts().put(2, o5);
+                doc3.getContacts().put(2, o6);
 
                 doc1.getContacts2().put(2, o1);
                 doc2.getContacts2().put(2, o2);
@@ -79,6 +95,9 @@ public class EntityViewSettingTest extends AbstractEntityViewTest {
                 em.persist(o1);
                 em.persist(o2);
                 em.persist(o3);
+                em.persist(o4);
+                em.persist(o5);
+                em.persist(o6);
 
                 em.persist(doc1);
                 em.persist(doc2);
@@ -114,7 +133,7 @@ public class EntityViewSettingTest extends AbstractEntityViewTest {
 
     @Test
     public void testEntityViewSettingWithEntityAttribute() {
-        EntityViewManager evm = build(DocumentWithEntityView.class);
+        EntityViewManager evm = build(DocumentWithEntityView.class, PersonView.class);
 
         // Base setting
         EntityViewSetting<DocumentWithEntityView, PaginatedCriteriaBuilder<DocumentWithEntityView>> setting = EntityViewSetting
@@ -135,7 +154,7 @@ public class EntityViewSettingTest extends AbstractEntityViewTest {
     
     @Test
     public void testEntityViewSettingNotExistingFilterAttribute() {
-        EntityViewManager evm = build(DocumentWithEntityView.class);
+        EntityViewManager evm = build(DocumentWithEntityView.class, PersonView.class);
 
         // Base setting
         EntityViewSetting<DocumentWithEntityView, PaginatedCriteriaBuilder<DocumentWithEntityView>> setting = EntityViewSetting
@@ -177,7 +196,7 @@ public class EntityViewSettingTest extends AbstractEntityViewTest {
 
     @Test
     public void testEntityViewFetches() {
-        EntityViewManager evm = build(DocumentWithEntityView.class);
+        EntityViewManager evm = build(DocumentWithEntityView.class, PersonView.class);
 
         EntityViewSetting<DocumentWithEntityView, CriteriaBuilder<DocumentWithEntityView>> setting = EntityViewSetting.create(DocumentWithEntityView.class);
         setting.fetch("id");
@@ -187,5 +206,34 @@ public class EntityViewSettingTest extends AbstractEntityViewTest {
         assertEquals("MyTest", view.getName());
         assertNotNull(view.getId());
         assertNull(view.getOwner());
+    }
+
+    @Test
+    @Category({ NoMySQLOld.class, NoHibernate42.class, NoHibernate43.class, NoHibernate50.class, NoHibernate51.class, NoEclipselink.class, NoDatanucleus.class, NoOpenJPA.class })
+    // We need a left entity join for this so Hibernate < 5.1 can't be used
+    // MySQL before 8 didn't support lateral and also don't support correlated LIMIT subqueries in quantified predicates
+    // EclipseLink doesn't support subqueries in functions which is required for LIMIT
+    // Datanucleus fails because of a NPE?
+    // OpenJPA has no function support
+    public void testEntityViewFetchesWithFilterAndSorter() {
+        EntityViewManager evm = build(DocumentWithEntityView.class, PersonView.class);
+
+        EntityViewSetting<DocumentWithEntityView, CriteriaBuilder<DocumentWithEntityView>> setting = EntityViewSetting.create(DocumentWithEntityView.class);
+        setting.fetch("id");
+        setting.addAttributeSorter("name", Sorters.ascending());
+        setting.addAttributeFilter("owner", false);
+        setting.addAttributeSorter("firstContact.name", Sorters.ascending());
+        setting.addAttributeSorter("contactCount", Sorters.ascending());
+
+        List<DocumentWithEntityView> list = evm.applySetting(setting, cbf.create(em, Document.class)).getResultList();
+        assertEquals(3, list.size());
+        DocumentWithEntityView view = list.get(0);
+        assertNotNull(view.getId());
+        assertNull(view.getName());
+        assertNull(view.getOwner());
+        // We need to fetch it to provide sorting
+        assertNotNull(view.getFirstContact());
+        Document document = cbf.create(em, Document.class).where("name").eq("MyTest").getSingleResult();
+        assertEquals(document.getId(), view.getId());
     }
 }